refactor: extract search item to separate component
